Adjusting to Google-style I
