+++ src/main/java/mops/portfolios/HardMock.java
@@ -1,9 +1,9 @@
 package mops.portfolios;
 
-import mops.portfolios.Domain.Entry.Entry;
-import mops.portfolios.Domain.Entry.EntryField;
-import mops.portfolios.Domain.Portfolio.Portfolio;
-import mops.portfolios.Domain.UserGroup.User;
+import mops.portfolios.domain.entry.Entry;
+import mops.portfolios.domain.entry.EntryField;
+import mops.portfolios.domain.Portfolio.Portfolio;
+import mops.portfolios.domain.UserGroup.User;
 import org.springframework.stereotype.Service;
 
 import java.util.Arrays;
