+++ src/main/java/mops/portfolios/PortfoliosController.java
@@ -1,8 +1,13 @@
 package mops.portfolios;
 
+import java.io.IOException;
+import java.nio.charset.StandardCharsets;
+import java.util.List;
+import javax.annotation.security.RolesAllowed;
+import javax.servlet.http.HttpServletRequest;
 import lombok.AllArgsConstructor;
-import mops.portfolios.Domain.Entry.Entry;
-import mops.portfolios.Domain.Portfolio.Portfolio;
+import mops.portfolios.domain.entry.Entry;
+import mops.portfolios.domain.Portfolio.Portfolio;
 import mops.portfolios.keycloak.Account;
 import mops.portfolios.tools.AsciiDocConverter;
 import org.keycloak.KeycloakPrincipal;
@@ -13,12 +18,6 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.multipart.MultipartFile;
-
-import javax.annotation.security.RolesAllowed;
-import javax.servlet.http.HttpServletRequest;
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
-import java.util.List;
 
 @Controller
 @AllArgsConstructor
@@ -142,8 +141,9 @@ public class PortfoliosController {
   @GetMapping("/portfolio")
   @RolesAllowed({"ROLE_orga", "ROLE_studentin"})
 
-  public String clickPortfolio(Model model, @RequestParam String title, KeycloakAuthenticationToken token) {
-  
+  public String clickPortfolio(Model model, @RequestParam String title,
+                               KeycloakAuthenticationToken token) {
+
     authorize(model, token);
   
     Portfolio portfolio = hardMock.getPortfolioByTitle(title);
@@ -161,20 +161,22 @@ public class PortfoliosController {
   }
 
   /**
-   * Entry mapping for GET requests.
+   * entry mapping for GET requests.
    *
    * @param model The spring model to add the attributes to
    * @param title The name of the entry
-   * @param entry_title The title of the entry
+   * @param entryTitle The title of the entry
    * @return The page to load
    */
   @SuppressWarnings("PMD")
   @GetMapping("/entry")
   @RolesAllowed({"ROLE_orga", "ROLE_studentin"})
-  public String clickEntry(Model model, @RequestParam String title, @RequestParam String entry_title, KeycloakAuthenticationToken token) {
+  public String clickEntry(Model model, @RequestParam String title,
+                           @RequestParam String entryTitle, KeycloakAuthenticationToken token) {
+
     authorize(model, token);
     Portfolio portfolio = hardMock.getPortfolioByTitle(title);
-    Entry entry = hardMock.getEntryByTitle(entry_title);
+    Entry entry = hardMock.getEntryByTitle(entryTitle);
 
     model.addAttribute("portfolio", portfolio);
     model.addAttribute("entry", entry);
@@ -221,7 +223,8 @@ public class PortfoliosController {
   @SuppressWarnings("PMD")
   @PostMapping("/view")
   @RolesAllowed({"ROLE_orga"})
-  public String viewUploadedTemplate(Model model, @RequestParam("file") MultipartFile file, KeycloakAuthenticationToken token) {
+  public String viewUploadedTemplate(Model model, @RequestParam("file") MultipartFile file,
+                                     KeycloakAuthenticationToken token) {
     authorize(model, token);
 
     byte[] fileBytes;
