+++ src/main/java/mops/portfolios/domain/Portfolio/PortfolioConfig.java
@@ -1,4 +1,4 @@
- package mops.portfolios.Domain.Portfolio;
+ package mops.portfolios.domain.Portfolio;
 
  import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
  import org.springframework.transaction.annotation.EnableTransactionManagement;
