+++ src/main/java/mops/portfolios/domain/UserGroup/UserGroup.java
@@ -1,4 +1,4 @@
-package mops.portfolios.Domain.UserGroup;
+package mops.portfolios.domain.UserGroup;
 
 import lombok.*;
