+++ src/main/java/mops/portfolios/domain/UserGroup/UserGroupRepository.java
@@ -1,4 +1,4 @@
-package mops.portfolios.Domain.UserGroup;
+package mops.portfolios.domain.UserGroup;
 
 import org.springframework.data.repository.CrudRepository;
