+++ src/main/java/mops/portfolios/domain/entry/EntryRepository.java
@@ -1,4 +1,4 @@
-package mops.portfolios.Domain.Entry;
+package mops.portfolios.domain.entry;
 
 import org.springframework.data.repository.CrudRepository;
